03_API_service/backend/requirements.txt
# Flask 核心框架
Flask==3.0.0
Flask-CORS==4.0.0
# 数据库相关
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
# JWT 认证
Flask-JWT-Extended==4.6.0
# 数据处理
pandas==2.1.4
numpy==1.26.2
# 机器学习（用于加载模型）
scikit-learn==1.3.2
tensorflow==2.15.0
# 工具库
python-dotenv==1.0.0